test: move Python code from strings to separate files
