+++ tests/test_cli_model.py
@@ -5,54 +5,8 @@ import pytest
 pytest_plugins = ["pytester"]
 
 CONFTEST = (Path(__file__).parent / "conftest.py").read_text()
-TEST_ALREADY_EXISTS = """
-from typing import Any
-
-import pytest
-import pytest_jubilant
-import jubilant
-
-
-def _mock_add_model(self, model: str, *args: Any, **kwargs: Any):
-    msg = (
-        f'ERROR failed to create new model: '
-        f'model "{model}" for admin already exists (already exists)'
-    )
-    raise jubilant.CLIError(1, ["juju", "add-model"], stderr=msg)
-
-
-@pytest.fixture(scope="module", autouse=True)
-def _patch_add_model():
-    with pytest.MonkeyPatch.context() as monkeypatch:
-        monkeypatch.setattr(jubilant.Juju, "add_model", _mock_add_model)
-        yield
-
-
-def test_create_model(temp_model_factory: pytest_jubilant.TempModelFactory):
-    temp_model_factory.get_juju("my-fancy-model")
-""".strip()
-TEST_OTHER_CLI_ERROR = """
-from typing import Any
-
-import pytest
-import pytest_jubilant
-import jubilant
-
-
-def _mock_add_model(self, model: str, *args: Any, **kwargs: Any):
-    raise jubilant.CLIError(1, ["juju", "add-model"], stderr="ERROR something else")
-
-
-@pytest.fixture(scope="module", autouse=True)
-def _patch_add_model():
-    with pytest.MonkeyPatch.context() as monkeypatch:
-        monkeypatch.setattr(jubilant.Juju, "add_model", _mock_add_model)
-        yield
-
-
-def test_create_model(temp_model_factory: pytest_jubilant.TempModelFactory):
-    temp_model_factory.get_juju("my-fancy-model")
-""".strip()
+TEST_ALREADY_EXISTS = (Path(__file__).parent / "cli_model_tests_already_exists.py").read_text()
+TEST_OTHER_CLI_ERROR = (Path(__file__).parent / "cli_model_tests_other_cli_error.py").read_text()
 
 
 def test_explicit_model_allows_collisions(pytester: pytest.Pytester):
